requirements.txt
awscrt
awsiot
python-dotenv